backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
google-genai>=0.3.0
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
PyPokerEngine
